# [process][posix] Fix #1049 check if procfs is mounted before checking if pid exists there